Document how to access UART on BliKVM v4
This change modifies the BliKVM-v4-guide and adds a section on how to
use the CH341 serial adapter built into BliKVM v4.

The presence of a USB to UART adapter is not documented in the dev
guide for BliKVM v4 and it may seem like the USB data pins need to
be attached to a UART on the other side - this is not the case and the
external UART is not needed.